# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/v1/users": {
      "get": {
        "operationId": "UsersAdminController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 9007199254740991,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List staff accounts for this tenant (paginated)",
        "tags": [
          "Admin / Users"
        ]
      },
      "post": {
        "operationId": "UsersAdminController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create a new admin or staff account",
        "tags": [
          "Admin / Users"
        ]
      }
    },
    "/admin/v1/users/{id}/role": {
      "patch": {
        "operationId": "UsersAdminController_changeRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Change the role of a staff account (admin/staff only; not owner; not self)",
        "tags": [
          "Admin / Users"
        ]
      }
    },
    "/admin/v1/users/{id}/deactivate": {
      "patch": {
        "operationId": "UsersAdminController_deactivate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Deactivate a staff account (sets disabled_at; invalidates sessions)",
        "tags": [
          "Admin / Users"
        ]
      }
    },
    "/admin/v1/users/{id}/reactivate": {
      "patch": {
        "operationId": "UsersAdminController_reactivate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Reactivate a disabled staff account (clears disabled_at)",
        "tags": [
          "Admin / Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "SovEcom API",
    "description": "The SovEcom headless ecommerce API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 320,
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$",
            "x-nestjs_zod-parent-additional-properties": false
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "staff"
            ],
            "x-nestjs_zod-parent-additional-properties": false
          },
          "password": {
            "type": "string",
            "minLength": 12,
            "maxLength": 1024,
            "x-nestjs_zod-parent-additional-properties": false
          }
        },
        "required": [
          "email",
          "name",
          "role",
          "password"
        ]
      },
      "ChangeRoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "staff"
            ],
            "x-nestjs_zod-parent-additional-properties": false
          }
        },
        "required": [
          "role"
        ]
      },
...
    }
  }
}